+++ shared/src/iosMain/kotlin/eu/heha/conifer/IosConiferApp.kt
@@ -16,7 +16,7 @@ object IosConiferApp {
             reportShareController = IosReportShareController,
             logFileInitializer = IosLogFileInitializer,
             uncaughtErrorInitializer = IosUncaughtErrorInitializer,
-            logClosingInitializer = IosLogClosingInitializer
+            appPresenceInitializer = IosAppPresenceInitializer
         )
     }
 }